requirements.txt
gradio
scikit-learn
numpy
sklearn.preprocessing
seaborn
matplotlib